+++ FindPanel/OgreTextFindThreads/OgreReplaceAllThread.m
@@ -54,7 +54,7 @@
         return;
     }
     
-    unsigned    stringLength = [string length];
+    NSUInteger  stringLength = [string length];
     
     NSRange     selectedRange = [aLeaf selectedRange];
 	if (![self inSelection]) {
@@ -151,7 +151,7 @@
     cancelledMessagePlural      = OgreTextFinderLocalizedString(@"%d strings replaced. (canceled, %.3fsec)");
     
     NSString    *message;
-    unsigned    count = [self numberOfMatches];
+    NSUInteger  count = [self numberOfMatches];
 	if ([self isTerminated]) {
 		if (count == 0) {
 			NSBeep();
